Use cluster.scheduler_info rather than cluster.scheduler
Supercedes https://github.com/dask/dask-labextension/pull/72

This depends on https://github.com/dask/distributed/pull/2902 , which
adds a `Cluster.scheduler_info` attribute to clusters which holds
necessary scheduler information.  We prefer this over querying a
Scheduler object directly in case that scheduler is not local, as in
increasingly becoming the case.